database/sinkdb: add Close method
Add Close method to release resources associated with a sinkdb handle
and stop its raft service. Also use this Close method to write a basic
unit
test for persistence.

Ideally, we could unit test without gracefully closing the database and
the
raft WAL, but the WAL file lock won't be released until the process exits
and the file descriptors are released.

Closes #1342