docker/ci: Update Rust to nightly-2017-06-15
This includes some pretty major upstream changes that affect us: libcollections
was merged into liballoc, and that's where we get all of our heap-allocated data
structures.

Also bumps rustfmt to 0.9.0, which includes a number of changes to the rustfmt
style guide.

Lastly, bumps clippy to 0.0.140, which is confirmed to build on this nightly
release.