dashboard: fix polyfill configuration for generators
We want to experiment with the new async/await syntax in the 
future. This corrects an ordering issue in the polyfill loading 
that prevented generators (and hence async/await) from running.

New syntax is allowed in eslint, but no transform plugin is provided
at this time. I recommend the following for future use:
https://github.com/babel/babel/tree/master/packages/babel-plugin-transform-async-to-generator

Closes #1352