net/raft: add unit test for node restart
Adds unit tests and a fix for restarting a node in a cluster. Previously,
restarting the initial node in a cluster would segfault since there was no
snapshot to restore from. 

Now, we trigger snapshot while initializing a raft cluster to ensure that
we can always restart nodes without errors.

Closes #1381