core/config: add Options type
Add a new Options type for storing and accessing incremental
configuration options. For now, this only implements sets because that's
what we have an immediate need for. Scalar values will come later.

DefineSet defines a new configuration option that is set of k-tuples.
The caller must pass in a CleanFunc, used to validate and canonicalize
user input, and a EqualFunc, used to determine equality between tuples.

This PR only adds the Options type. Follow-ups will add the RPC handler
logic, corresponding corectl commands and the first config option.

Closes #1384