core/config: update tuple on Add
For set configuration options, equality may be defined on a subset of
the fields of the tuple.

For example, for HSM URLs, we'll store sets of (URL, token) tuples
with equality defined on the URL. Previously, if we saw a tuple that
matched the key, we would perform a no-op since the tuple already exists.
Instead, we should update the existing tuple so that the non-key portions
of the tuple are updated.

```
corectl add hsmurls https://hsm1.chain.com my-access-token
corectl add hsmurls https://hsm1.chain.com real-token
corectl get hsmurls
```