core/config: introduce AddOrUpdate
For set configuration options, equality may be defined on a subset of
the fields of the tuple. Previously, we treated an Add of a tuple with
an existing key as a no-op. This splits the Add method into Add and
AddOrUpdate. Add will error if a tuple's key conflicts with an existing
tuple that is not exactly equal to the new tuple. AddOrUpdate will
replace the conflicting tuple.

This commit also updates corectl to allow providing the `-u` flag to
use the AddOrUpdate behavior.

Closes #1395