core: delete access token grants in a batch
When deleting an access token, delete all of the associated
grants in a single batch. Previously, we'd delete the grants for
each policy separately.

If _n_ is the number of policies, we'd need to wait for _2n_ raft
consensus rounds to delete all of the grants (1 for a linearizable
read, 1 for the write). Now we wait _n+1_ consensus rounds. Since
grants.Delete is just doing a read-modify-write, we should be able
to do an optimistic stale read so we only have to wait for a
single consensus round.

Since this change returns a sinkdb.Op, it'll also make it easier
to delete access tokens and their grants atomically down the
road.

Closes #1410