dashboard: generate licenses during production builds
This commit adds `license-webpack-plugin` to the dashboard build pipeline.
A new file, `3rdpartylicenses.txt`, will be generated in our production
webpack bundle. The primary use case for this file is to as the basis for a
more accurate NOTICE file in our dashboard source directory, one that only
includes licenses for dependencies that we are actually re-distributing.

Closes #1424