database/localdb: new package
This adds a new package, localdb, for writing to RocksDB. 

For now, this package exposes a single write operation, `Put`. We will soon
want to replace this with something that supports test-and-sets, perhaps
like `Exec` from sinkdb. However, that likely will require more design, so
let's start here, for now.  

Closes #1411