core/config: add single-tuple config options
Add corectl set command.

As-is, GetFunc provides no distinction between values that are not set and
values that are set to the empty byte slice. I'm unsure if we have any need
for that distinction or how we should expose it if we do.

Also, I stored the tuples in the same format as sets (a ValueSet protobuf),
so that all config options can be accessed through List regardless of their
type.

Closes #1446