docs: remove Windows installer (#1474)
We're a small team, focused on making [Sequence](https://seq.com)
a great product over the last few months. We think Sequence is a going
to be a great upgrade path for many Chain Core Developer Edition users.

In the meantime, we're continuing to support Chain Core DE in the
[Community Slack account](https://chain.slack.com/) and elsewhere.

However, we feel we can no longer provide great support specifically for
the Windows installer for Chain Core DE.

This change removes the instructions for using that installer, leaving
Docker as the preferred installation method on Windows.